stub: move runBlockValidation host operation onto StubService, leave a thin endpoint
